bot/requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.5.1
distlib==0.3.1
fastapi==0.61.2
filelock==3.0.12
h11==0.11.0
idna==2.10
multidict==4.7.6
pydantic==1.7.2
python-dotenv==0.15.0
six==1.15.0
starlette==0.13.6
uvicorn==0.12.2
virtualenv==20.1.0
yarl==1.5.1